# doc/openapi.yaml
openapi: 3.0.0
info:
  title: Playzone API
  version: '0.1'
servers:
  - url: 'http://localhost:8000'
paths:
...
  /api/projects:
    parameters: []
    get:
      summary: Récupérer des projets
      tags:
        - project
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  pages:
                    type: integer
                    minimum: 0
                  page:
                    type: integer
                    minimum: 0
                  size:
                    type: integer
                    minimum: 1
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForm'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
      operationId: get-api-projects
      security:
        - ADMIN: []
        - USER: []
      parameters:
        - schema:
            type: integer
            minimum: 0
          in: query
          name: page
          description: index de la pagination
          required: true
        - schema:
            type: integer
            minimum: 1
          in: query
          name: size
          description: taille de la pagination
          required: true
        - schema:
            type: string
            enum:
              - OWNER
              - SHARED
              - PUBLIC
          in: query
          name: from
          description: 'pour n''afficher que les projets créés par l''utilisateur connecté, ou les projets partagés avec l''utilisateur connecté, ou les projets publics de tous les utilisateurs, sinon affiche tous les projets si ce champ est omis et que l''utilisateur connecté est un administrateur'
    post:
      summary: Créer un projet
      operationId: post-api-projects-projectId
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema:
                type: string
              description: '/api/projects/{projectId}'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Created'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForm'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
      security:
        - ADMIN: []
        - USER: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectFormCreate'
      tags:
        - project
  '/api/projects/{projectId}':
    parameters:
      - schema:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
        name: projectId
        in: path
        required: true
        description: id du projet
    get:
      summary: Récupérer un projet
      tags:
        - project
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
        '404':
          description: Not Found
          content: {}
      operationId: get-api-projects-projectId
      security:
        - ADMIN: []
        - USER: []
    put:
      summary: Modifier un projet
      operationId: put-api-projects-projectId
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForm'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
        '404':
          description: Not Found
          content: {}
      security:
        - ADMIN: []
        - USER: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectFormUpdate'
      tags:
        - project
    delete:
      summary: Supprimer un projet
      operationId: delete-api-projects-projectId
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
        '404':
          description: Not Found
          content: {}
      security:
        - ADMIN: []
        - USER: []
      tags:
        - project
...
  '/api/project/{projectId}/permissions':
    parameters:
      - schema:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
        name: projectId
        in: path
        required: true
        description: id du projet
    get:
      summary: Récupérer les permissions d'accès à un projet
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectPermissions'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
        '404':
          description: Not Found
          content: {}
      operationId: get-api-projects-projectId-permissions
      security:
        - ADMIN: []
        - USER: []
      tags:
        - permission
  '/api/project/{projectId}/permissions/{userEmailBase64}':
    parameters:
      - schema:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
        name: projectId
        in: path
        required: true
        description: id du projet
      - schema:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
        name: userEmailBase64
        in: path
        required: true
        description: email en base64 de l'utilisateur avec qui partager le projet
    get:
      summary: Récupérer une permission d'accès à un projet
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectPermission'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
        '404':
          description: Not Found
          content: {}
      operationId: get-api-projects-id-permissions
      security:
        - ADMIN: []
        - USER: []
      tags:
        - permission
    post:
      summary: Créer ou modifier une permission d'accès à un projet
      operationId: put-api-projects-id-permissions
      responses:
        '204':
          description: No Content
          headers: {}
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForm'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
      security:
        - ADMIN: []
        - USER: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectPermission'
      tags:
        - permission
    delete:
      summary: Supprimer une permission d'accès à un projet
      operationId: delete-api-projects-id-permissions
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
        '404':
          description: Not Found
          content: {}
      security:
        - ADMIN: []
        - USER: []
      tags:
        - permission
...
components:
  schemas:
...
    Project:
      title: Project
      type: object
      properties:
        id:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
        name:
          type: string
          example: Projet Emplois 2023
        description:
          type: string
          example: Statistiques de l'emploi en 2023
        isPublic:
          type: boolean
        log:
          $ref: '#/components/schemas/HistoryLog'
      required:
        - id
        - name
        - description
        - isPublic
        - log
    ProjectFormCreate:
      title: ProjectFormCreate
      type: object
      properties:
        name:
          type: string
          example: Projet Emplois 2023
        description:
          type: string
          example: Statistiques de l'emploi en 2023
      required:
        - name
        - description
    ProjectFormUpdate:
      title: ProjectFormUpdate
      type: object
      properties:
        name:
          type: string
          example: Projet Emplois 2023
        description:
          type: string
          example: Statistiques de l'emploi en 2023
        isPublic:
          type: boolean
      required:
        - name
        - description
        - isPublic
    ProjectPermission:
      title: ProjectPermission
      type: object
    ProjectPermissions:
      title: ProjectPermissions
      type: array
      items:
        type: object
        properties:
          email:
            type: string
            format: email
        required:
          - email
...
    ErrorForm:
      title: ErrorForm
      type: object
      properties:
        status:
          type: integer
          example: 400
        path:
          type: string
          example: /api/endpoint
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: name
              message:
                type: string
                example: required
            required:
              - field
              - message
      required:
        - status
        - path
        - errors
...
    ErrorUnauthorized:
      title: ErrorUnauthorized
      type: object
      properties:
        status:
          type: integer
          example: 401
        path:
          type: string
          example: /api/endpoint
        error:
          type: string
          example: Unauthorized
      required:
        - status
        - path
        - error
    ErrorForbidden:
      title: ErrorForbidden
      type: object
      properties:
        status:
          type: integer
          example: 403
        path:
          type: string
          example: /api/endpoint
        error:
          type: string
          example: Forbidden
      required:
        - status
        - path
        - error
    Created:
      title: Created
      type: object
      properties:
        id:
          type: string
          example: 01BX5ZZKBKACTAV9WEVGEMMVRY
      required:
        - id
    HistoryLog:
      title: HistoryLog
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: email
        updatedBy:
          type: string
          format: email
      required:
        - createdAt
        - updatedAt
        - createdBy
        - updatedBy
  securitySchemes:
    ADMIN:
      type: http
      scheme: bearer
      description: ''
    USER:
      type: http
      scheme: bearer
...